[setup] Only install rustup if rust-analyzer is installed
This will skip rustup on my work computer.

+++ setup.sh
@@ -20,9 +20,11 @@ cd ~/.dotfiles.public
 brew bundle install
 
 # Rust
-rustup --version || curl --proto '=https' --tlsv1.2 -sSf https://sh.rustup.rs | bash -s -- --no-modify-path
-rustup toolchain install stable
-rustup component add rust-src rustfmt clippy
+if [[ -f "$(brew --prefix)/bin/rust-analyzer" ]]; then
+  rustup --version || curl --proto '=https' --tlsv1.2 -sSf https://sh.rustup.rs | bash -s -- --no-modify-path
+  rustup toolchain install stable
+  rustup component add rust-src rustfmt clippy
+fi
 
 # Python shivs and other things
 make -j$(nproc)
